Move EVM-related functions into its own directory
