+++ src/components/SendTokens.vue
@@ -14,8 +14,8 @@ import { useActivitiesStore } from '@/store/activities'
 import { useRpcStore } from '@/store/rpc'
 import { useUserStore } from '@/store/user'
 import { getTokenBalance } from '@/utils/contractUtil'
+import { getRequestHandler } from '@/utils/evm/requestHandlerSingleton'
 import { convertGweiToEth } from '@/utils/gweiToEth'
-import { getRequestHandler } from '@/utils/requestHandlerSingleton'
 import { getStorage } from '@/utils/storageWrapper'
 import { truncateToTwoDecimals } from '@/utils/truncateToTwoDecimal'
 import { useImage } from '@/utils/useImage'
