+++ src/pages/loggedInView.vue
@@ -14,21 +14,21 @@ import { useParentConnectionStore } from '@/store/parentConnection'
 import { useRequestStore } from '@/store/request'
 import { useRpcStore } from '@/store/rpc'
 import { useUserStore } from '@/store/user'
-import { AccountHandler } from '@/utils/accountHandler'
 import { GATEWAY_URL, AUTH_NETWORK } from '@/utils/constants'
 import { createParentConnection } from '@/utils/createParentConnection'
-import { getAuthProvider } from '@/utils/getAuthProvider'
-import getValidAppMode from '@/utils/getValidAppMode'
-import { getWalletType } from '@/utils/getwalletType'
+import { EthereumAccountHandler } from '@/utils/evm/ethereumAccountHandler'
 import {
   getRequestHandler,
   setRequestHandler,
-} from '@/utils/requestHandlerSingleton'
+} from '@/utils/evm/requestHandlerSingleton'
 import {
   getSendRequestFn,
   handleRequest,
   watchRequestQueue,
-} from '@/utils/requestManagement'
+} from '@/utils/evm/requestManagement'
+import { getAuthProvider } from '@/utils/getAuthProvider'
+import getValidAppMode from '@/utils/getValidAppMode'
+import { getWalletType } from '@/utils/getwalletType'
 import { getStorage } from '@/utils/storageWrapper'
 
 const userStore = useUserStore()
@@ -52,13 +52,13 @@ onBeforeMount(() => {
 
 onMounted(async () => {
   setRpcConfigs()
-  initKeeper()
+  await initKeeper()
   connectToParent()
   await setTheme()
   await getRpcConfig()
   await getAccountDetails()
   await setMFABannerState()
-  router.push({ name: 'home' })
+  await router.push({ name: 'home' })
   loader.value.show = false
 })
 
@@ -95,7 +95,10 @@ async function getAccountDetails() {
 }
 
 async function initKeeper() {
-  const accountHandler = new AccountHandler(
+  const pc = await parentConnection.promise
+  const cfg = await pc.getAppConfig()
+
+  const accountHandler = new EthereumAccountHandler(
     userStore.privateKey,
     rpcStore.selectedRpcConfig.rpcUrls[0]
   )
